# Conflicts:
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/dao/PipelineBuildHistoryDao.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/dao/PipelineBuildStageDao.kt

+++ src/backend/ci/core/dispatch/api-dispatch/src/main/kotlin/com/tencent/devops/dispatch/pojo/PipelineBuildCreate.kt
@@ -31,5 +31,5 @@ data class PipelineBuildCreate(
     val pipelineId: String,
     val buildId: String,
     val vmSeqId: String,
-    val vmId: Int
+    val vmId: Long
 )
